Add descriptors module
