+++ src/boxes.rs
@@ -5,9 +5,9 @@ use std::{
 };
 
 use crate::{
-    basic_types::as_box_object, io::ExternalBytes, BaseBox, BoxHeader, BoxSize, BoxType, Decode,
-    Either, Encode, Error, FixedPointNumber, FullBox, FullBoxFlags, FullBoxHeader, Mp4FileTime,
-    Result, Uint, Utf8String,
+    basic_types::as_box_object, descriptors::SlConfigDescriptor, io::ExternalBytes, BaseBox,
+    BoxHeader, BoxSize, BoxType, Decode, Either, Encode, Error, FixedPointNumber, FullBox,
+    FullBoxFlags, FullBoxHeader, Mp4FileTime, Result, Uint, Utf8String,
 };
 
 /// ペイロードの解釈方法が不明なボックスを保持するための構造体
@@ -4216,41 +4216,6 @@ impl Decode for DecoderSpecificInfo {
 
         let mut payload = vec![0; size];
         reader.read_exact(&mut payload)?;
-
-        Ok(Self { payload })
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-#[allow(missing_docs)]
-pub struct SlConfigDescriptor {
-    pub payload: Vec<u8>,
-}
-
-impl Decode for SlConfigDescriptor {
-    fn decode<R: Read>(mut reader: R) -> Result<Self> {
-        let tag = u8::decode(&mut reader)?;
-        if tag != 6 {
-            // 6 = SLConfigDescrTag
-            return Err(Error::invalid_data(&format!(
-                "Unexpected descriptor tag: expected=6, actual={tag}"
-            )));
-        }
-
-        // TODO:
-        let mut size = 0;
-        let mut has_next_byte = true;
-        while has_next_byte {
-            let b = u8::decode(&mut reader)?;
-            has_next_byte = Uint::<u8, 1, 7>::from_bits(b).get() == 1;
-            size = (size << 7) | Uint::<u8, 7>::from_bits(b).get() as usize
-        }
-        dbg!(size);
-
-        // TODO:
-        let mut payload = vec![0; size];
-        reader.read_exact(&mut payload)?;
-        dbg!(&payload);
 
         Ok(Self { payload })
     }
